requirements.txt
Flask
RPi.GPIO; platform_system == "Linux"
